views: split events_register.py from main views file.
